Clear events from voting while awaiting runtime update
Voting events have been introduced in #4613 and #6544, but Polkadot/Kusama
runtimes have not yet been updated to reflect this.

Instead of storing ALL events for the transactions containing the voting
extrinsics, only those for `convictionVoting` are stored i.e. they will
be empty, which will trigger an error anyway when the runtimes are
updated.