Correct proxy filtering tests (#266)
* Rework proxy action builder

`concat()` was used instead of `push()` in some proxy action builders,
which means some proxy call filtering tests were running on empty lists
of actions.

This commit fixes the case for allowed call tests.

* Update snapshots

* Fix `ParaRegistration` filtering tests

The reason `remove_proxy` was not working was that the proxy type passed
to the extrinsic was hardcoded, when in reality its index is different
on Polkadot and Kusama; this is combined with the fact the that proxy
type can only remove itself.

* Improve `TODO`s on proxy filtering test module