Create E2E test suite for accounts and balances (#377)
* Init balances e2e tests (wip)

* Rework initial tests for clarity

Most of them were written with recourse to Cursor, and were just not
very useful.

* Refactor `transfer_allow_death` test

It scrutinizes the extrinsic's events and their inner data.

* Recreate tests for all Polkadot/Kusama relay/SPs

* Rework networks that run `transferAllowDeath` test

Now, at the calling site of the test tree creation function, chains can
signal whether their ED is too low to run the basic `transferAllowDeath`
test.

* Scrutinize more events in `transferAllowDeath` test

* Test non-destructive `transfer_allow_death`

* Test `force_transfer` with account reaping

* Test to `force_transfer` reaping

* Test `transfer_allow_death` below ED

* Test `force_transfer` below ED

* Test overtransferring account

* Update comments and snapshots

* Test `force_transfer` with insufficient funds

* Test self `transfer_allow_death` of near-entire balance

* Update snapshots

* Update accounts tests after merge from `master`

PR #384 required some additional fixes that auto-merge could not perform.

* Update snapshots

* Fund new test account on Bridge Hub

* Revert change to Polkadot AH multisig suite name

* Create test to liquidity restrictions bug

See https://github.com/paritytech/polkadot-sdk/pull/9560/ and
https://github.com/paritytech/polkadot-sdk/pull/8108.

TL;DR proxy and multisig pallets are using the old `Currency` trait
and not the new fungible traits, it is possible for an operation that
is permissible to an account at a given point in time and state to
fail.

See https://github.com/open-web3-stack/polkadot-ecosystem-tests/issues/401

* Test liquidity restrictions when creating proxy

* Parametrize deposit-requiring action liquidity restriction test

The `balances.LiquidityRestrictions` error can occur is different
contexts, as multiple pallets still use the old `Currency::reserve`
operation.

This commit parametrizes the test to allow for different kinds of
actions to be tested, not just proxy addition.

* Extend liquidity restriction tests to cover more cases

The test now combinatorially covers several cases:
1. Reserve action can vary between staking bond and nomination pool creation
2. Lock action is always vested transfer (for now)
3. Deposit action can vary between proxy creation, referendum submission
   and staking bond

* Test manual reserve/lock in liquidity restriction tests

This adds the possibility of manually set reserve/locks to the reserve/
lock actions used to combinatorially generate tests to liquidity
restrictions.

This is needed because some chains have no vesting or staking, but the
behavior should still be tested.

* Fix manual reserve/lock actions in liq. restriction tests

Fees were not being accounted for correctly when performing checks
on free/frozen balances.

* Update snapshots and accounts E2E test tree call sites

* Filter vesting as locking action on Asset Hubs

... while the AHM is pending - vesting operations are filtered, and
thus cannot be used.

* Add more comments to liquidity restriction test

Especially the action interfaces.

* Add more comments to liquidity restriction tests

* Remove leftover debug code from self-transfer test

* Fix Asset Hub test suites' use of relay chain

An invalid argument was being passed to the `scheduleInlineCallWithOrigin`
function: the base's chains block provider, instead of the relay chain's,
which is always `Local`.

* Snapshot skipped liquidity restriction tests

... instead of logging a message. Logs add some noise to CI test output.

* Apply lint fixes

* Update block numbers

* Change lock identifier used

* Remove unstable `Transfer` event from snapshots

* Update snapshots

* Simplify liquidity restriction tests

The `Balance.locks` storage is no longer modified - Polkadot Relay is
failing for an undetermined reason.

* Use nonces when setting storage

* Test `transfer_allow_death` with reserve

* Add test to `transfer_all` with reserve

* Add `force_transfer` with reserve test

* Add origin check tests for gated extrinsics

* Add tests to transfer functions with insufficient funds

* Test `transfer_all` with `keepAlive = true`

* Correct previous transfer all test snapshot

* Test `transfer_all` with `keep_alive = false`

* Correct snapshots of accounts tests

* Test `transfer_keep_alive` with source left below ED

* Split `transfer_keep_alive` tests

... into two separate tests:
1. one for chains with low ED (below a typical transfer fee)
2. one for chains with normal ED

* Test `force_adjust_total_issuance` with zero delta

* Test total issuance changes

* Test `force_unreserve` no-op cases

* Update accounts E2E test snapshots

* Update liquidity restriction test snapshots for AHs

* Test forceful unreservation of funds

* Add more self-transfer tests

* Add self-transfer test to `force_transfer`

* Add tests to `force_set_balance`

* Refactor `burn` tests into low/normal ED tests

* Update snapshots with new `burn` tests

* Test burning from account with consumer reference

* Add test to burning of entire balance (or more than it)

* Update snapshots to remove obsolete data

* Improve documentation and `README`