Remove Kusama relaychain tests (#438)
* Bump block numbers

* Remove obsolete Kusama relay E2E tests

* Update snapshots for some Kusama tests (partial)

* Simplify action logic in liquidity restriction tests

* Update Kusama parachain liquidity E2E tests

Kusama parachains now include fix to liquidity restriction error.

* Update KAH accounts E2E snapshots

* Update PAH accounts E2E test tree

* Rename PAH proxy type object

* Bump block numbers again

* Update KAH proxy E2E tests

With the AHM, proxy types on AHM are now expanded to include the relay's
proxy types.

* Bump block numbers again

* Update Kusama proxy tests

Changes to other proxy tests in this commit are due to the renaming of
the proxy removal action for clarity.

* Update Kusama relay proxy test snapshots

* Update KAH XCM tests

Some are disabled pending a GH issue and subsequent PR.

* Update Kusama relay proxy test snapshots again

* Filter Kusama XCM tests temporarily

* Skip burn tests in Kusama relay

Temporary workaround; see
https://github.com/paritytech/polkadot-sdk/issues/9986

* Redact total issuance from events

This created some flakiness in tests.

* Update accounts tests snapshots